added http://code.google.com/p/sparsehash/ to deps for Options.cpp
on Linux, you may need to download the relevant .deb or .rpm package from that page!
using a hashtable to store a string map just makes sense

If sparsehash just does not compile on your system, try compiling with NO_GOOGLE_SPARSEHASH defined to fall back to std::map!

also added: a cache of bool and int options so they don't need to be re-parsed from text for every getInt() or getBool(); maybe be relevant in AI?